Include news.atom in the list of exceptions that don't go to xsl.py
xsl.py recieves all requests except those matching the pattern above
the "SetHandler None".  But it can only deal with those files in the
site's root dir ending in '.html'; for everything else it returns the
404 page.  '/news.atom' doesn't match the "SetHandler None" pattern,
so that's what happened to it.

In lshift-web, news.atom lives under downloads, so it matched the
"SetHandler None" pattern.

I've fixed the "SetHandler None" pattern so that it matches
/news.atom, /favicon.ico, and anything in the subdirectories under
site.