misc changes to ssl troubleshooting guide
The paths used in the examples never use mention the "path/to/" component.

Changed all the "if..then" structures to "if... ,", as the following Wikipedia
page suggests.  I'm not sure what the correct English is, but those sentences
look weird without the comma after the conditional clause.

    http://en.wikipedia.org/wiki/Conditional_sentence

Changed the last "Diagnostics" heading to a subsection, since that's what all
the other headings are.