update update clients.xml, documentation.xml, and site map
Documentation doesn't include a link to clients.xml anymore.  I left clients
alive, with a link to the edoc (which was missing) and the AMQP uri scheme.

I don't think it'll hurt to leave it there so that links won't break and we have
something to put in the site map.