Don't mention rabbitmq_clusterer
It is very opinionated and can introduce novel
issues in operations and system failure scenarios. Moreover,
it's reason for existence is now available in the core (as of 3.6.7+):
nodes can be restarted in arbitrary order with N retries every T
seconds.

Per discussion with the team.

Drive-by change: correct Chef cookbook link URL.