Make it easier to understand how cluster formation works
Service-specific details are pushed further down, the intro is all about
the big picture.

I would have done more re-shuffling, but I time-boxed it. This is a good
example for a doc that feels "jumbled up". There is either too much or
too little information, but never enough to form a clear &
straightforward picture of how the user can get from A to B.
Specifically:

  * The Erlang cookie is a detour and doesn't really belong here. Yes,
  it's important, but it would make more sense as an item in a list of
  pre-requisites

  * Imperative cluster configuration definitely does not belong here,
  not even as a mention. If I'm thinking Kubernetes or AWS auto-scaling
  groups, I do not care about hand-rolling clusters with rabbitmqctl

There are other things which I could mention, but the above should be
enough to maybe consider a shift in approach.