Improve memory use wording, remove repetitive links
A link is an invite to go elsewhere and can distract from the task at
hand. They are important, but, as with everything, it needs balancing. I
feel that removing some of the links is better, but happy to concede if
someone has a different, stronger opinion.

Some of these changes may be controversial, cc @michaelklishin