Federation: formatting, wording
While a description list is a reasonable choice semantically,
the way our guides are structured is with sections and subsections.
DLs are also formatted in a way that is sometimes counterproductive
with longer descriptions.

So let's switch to paragraphs and headings (primarily to fix the way things
are presented, I know, I know).

Note that DL abuse is quite common in the rest of federation guides which will
follow suit.