additional clarity for cluster_partition_handling
Start the chapter with the info, that a cluster might be used for different purposes and therefore has to be tuned accordingly.
Introduce the word split-brain and use it where appropriate for increased clarity (i.e. recovering from a split-brain).
Use datacenter instead of AZ (availability zone), but reference it to have a well known example of the expected network reliability.

I claim that "pause_minority" prevents split-brain during network partitions. In our tests this proofed correct. Please correct me, if this claim is wrong (let's keep the "suspend" an exception, as it is not really the network partition we try to protect against in HA setups).