Add a new troubleshooting page
This is the outcome of a short discussion with @murphe5 & @ansergit.
They were saying that a list of helpful commands when troubleshooting
RabbitMQ deployments would go a long way. There already is a
troubleshooting page, so maybe it would make sense to work this content
in that context. For now, I'm just capturing what was discussed, feel
free to take it in any direction that makes sense @murphe5 & @ansergit.

cc @MarcialRosales