Ubuntu and Debian guide: cover "unindexed" distributions
"unindexed" here refers to the state of the package
indexing on Bintray. Bionic can be used for distribution
name on any Ubuntu or  Debian release published
from 18.04 to this day (almost certainly including Ubuntu 19.10).

This is a limitation of our pipeline that requires
package testing on an "indexed" distribution and assumes
an AWS AMI exists for it. For Debian Buster, this is
not the case at the moment.

Erlang Debian packages use only a few most common
distribution names which complicates matters further.

Per discussion with @dumbbell.