Update tutorial-two-go: include code needed to run
When following the tutorial, we are asked to make some edits and then to run and test the changes, before been shown all the code needed. A call to the bodyFrom function is added, but not mentioned right until the bottom of the page. Including it here will help follow the tutorial as we go along. It will make line 144 actually work :-)