Delete more classic config examples
Note that in some cases, such examples
are required or really should be mentioned
because e.g. LDAP configuration will
unavoidably have certain bits of configuration
in the advanced.config file.

In other case, e.g. for basic TCP listener
configuration, such examples are nothing
but the relics of the past.