`no_ack` is actually turned _on_ by default
I'd rather prefer to use named parameters if it wasn't such a young feature in PHP (introduced just a month ago):
```php
$channel->basic_consume($queueName, no_ack: true, callback: $callback);
```
But since it's so young, maybe it's better to use an old syntax. But it's yet worth to fix the false statement that `no_ack` is turned off by default.