Update Monitoring RabbitMQ in Kubernetes
This commit adds docs on how to integrate RabbitMQ scrape targets,
alerting rules, and Grafana dashboards with an existing Prometheus.

We describe the integration for a Prometheus installed by Prometheus
Operator (i.e. CRDs ServiceMonitor, PodMonitor, PrometheusRule are available)
and for a Prometheus not installed by Prometheus Operator.

Remove the section to scrape RabbitMQ by annotations since that approach
is not recommended anymore.

This commit tries to remove duplicate YAML definitions from the docs
and instead links to the YAML definitions already present in the
cluster-operator repo.

Relates to https://github.com/rabbitmq/cluster-operator/pull/667 and
https://github.com/rabbitmq/cluster-operator/pull/676