Expand Cluster Sizing with more specific recommendations
I removed the link to HTTP API since it's highly unlikely that anyone
would use the HTTP API to determine which node to connect to. In the
production environments that I was involved with, there is usually a
load balancer in front of the RabbitMQ cluster, so this wouldn't even be
possible. Also, if an AMQP 0.9.1 client library is used, unless this
information is available in that context, I haven't seen a single
application use the HTTP API to dermine which RabbitMQ node to connect
to. Things are very different with Streams & the new protocol, and
hinting is definitely a step in the right direction.

That blog post is somewhat misleading because it benchmarks 186 queues
spread across 30 RAM nodes. Not only that we don't recommend clusters
that large, but the use of RAM nodes comes with many caveats. It is more
likely that users will have 3-node clusters with Quorum Queues, and
hopefully Streams not before long, so linking to that blog post is at
odds with the present and what is coming in 2021. Removing the link.

Is there anything that you want to add or refine @michaelklishin?

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>