Use heading levels & new lines for each heading level consistently
- h1 is at the top
- h2 is preceded by 3 new lines
- h3 is preceded by 2 new lines
- h4 is preceded by 1 new line

I think that we should limit ourselves to 3 heading levels at most, and
use bold (strong) text instead h4. If this doesn't feel like a good
change, feel free to drop it @michaelklishin. I captured this in a
separate commit that could be dropped in an interactive rewrite if you
think that is best.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>