Release calendar: Add 3.9.0-rc.2
A link is also added to the 3.9.0-rc.1 announcement.

While here rephrase "$version produced" to say "$version build starts".
The idea is that the build is started on the date mentionned, not
finished. It means that if a bug was not fixed or a patch was not merged
by that day, it will be taken care of *after* that specified version.