Remove initContainer override
since it's not needed anymore to deploy RabbitmqClusters on OpenShift
deployed by cluster-operator versions >= v1.9.0

due to the change in
https://github.com/rabbitmq/cluster-operator/commit/e4e4271134b0c5e33932a2749f3856e545d241b8#diff-543995f45618a67cfbfcc13ba8eab18b7b559478ff0d32b31456715d95240770L1104-L1106

Prior to that commit, RunAsUser was defined twice:
once in the PodSecurityContext, and once in the initContainer
SecurityContext, the latter overriding the former.
This was unnecessary since both SecurityContexts use the same user.
Therefore we removed the latter to reduce confusion and to make
deployment on OpenShift even easier.