Nuke ./how.html
This is a very old (literally more than one decade old) page that was
never updated.

We will set up a permanent redirect to /getstarted.html instead.

If at some point we decide to reintroduce such a "feature overview" page,
it should be throught through and written from scratch.

Per discussion with the team.