Updates to remove product-name; constant from topic titles
Hi @michaelklishin product-name; constant now removed from titles in topics. I have left in the title of the "migrate-mcq-to-qq.md" topic I wrote that myself so it has always been in it.

Also, can you please take a detailed look at the updated in "ttl.md" I have made quiet a few text changes in the opening section of that topic - many thanks