Update default exchange restrictions
RabbitMQ does not allow binding operations on the default exchange. Such
operations result in a 403 - Not Allowed error, at the time of this
writing.

We conform with the spec in the creation of the default exchange, and
default bindings i.e. every queue is implicitly bound to the default
exchange with routing key equal to queue name.

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>