Fix RABBITMQ_MNESIA_DIR description in relocate docs
While reading through the relocate docs, I noticed that a word seemed to
be missing from the `RABBITMQ_MNESIA_DIR` description.  A quick `git
grep` showed what the missing word should be from a similar description
in the configure docs.  This change makes the description for this
environment variable internally consistent.