Reorganize the docs table of contents
[Why]
The existing table of contents lack a logical sequence of topics and
looked more like a random succession of articles.

Furthermore, it didn’t take into account the two different profiles that
interact with the documentation: developers and administrators.

[How]
Topics that are not specifically about versions of the broker of nod
about the broker at all are moved outside of the `docs` folder. These
topics are:
* release information
* client libraries
* Kubernetes operator

With those out of the way, the ToC was reorganize in three main
sections:
1. How to install and upgrade
2. How to use (targetting developers)
3. How to manage (targetting admins)
4. How to monitor (also targetting admins, but important topic)