Wrangler: Prepare the deploy to v3-12.rabbitmq.com
[Why]
The current website will be replaced by a Docusaurus-base new website
that supports docs versioning. The new website will cover RabbitMQ
versions starting with 3.13.0.

We want to keep a copy of the docs of RabbitMQ 3.12.x. Therefore, we
want to publish what was the `live` branch to v3-12.rabbitmq.com.

[How]
The method remains the same: we use Cloudflare's Wrangler to deploy the
generated static website to a Cloudflare worker.