doc(tutorials): improve go instructions
Goal to improve the instructions by adding how to init a mod with Go, creating the main function and addind a note to create a new directory for the receiver. It is necessary because in GoLang we cannot have on the same directory, two files with the same name function. In this case, we cannot create two files with the same func main(){} name.