Cloudflare: Redirect categories manually
... instead of relying on the "subpath matching" feature.

But even with that, the redirect of categories doesn't seem to work. All
links point to www.rabbitmq.com/blog/tags. So the links are ok but
meaningless.

I'm still committing that update. But I'm almost sure it doesn't work. I
have zero confidence in this Cloudflare bulk redirects feature at this
point.