Rewrite homepage
[Why]
The homepage was old and presented outdated content.

It was in need for a heavy refresh to follow best practices of web
design and communication.

[How]
The new version is very inspired by the Rust homepage [1]. It has a hero
banner with the name of the project and a (new) tagline, plus a very
visible "Getting started" button.

It is followed by several sections describing RabbitMQ, giving examples
of use cases, telling about the license and sharing a couple user
feedbacks.

[1] https://www.rust-lang.org/