docs: Rewrite the introduction page
[Why]
The previous page only repeated the table of contents. Now that we have
a clean, reorganized and easy to access ToC in the left sidebar, we
don't need to duplicate the work.

[How]
The new introduction is now used to welcome users and:
- briefly explain the layout and organization of the docs
- guide them to the relevant content based on their profile, developer
  or administrator.

The introduction page also reminds the reader about the version of
RabbitMQ documentation they are currently viewing.