release-information: Rewrite the whole section
[Why]
The information was split between multiple pages and was also kind of
redundant with the release notes on GitHub.

It was time to clarify and make the content more readable.

[How]
The `changelog` and `news` pages are gone. They didn't bring much values
compared to what we had on GitHub. They barely served as announcements.
With RabbitMQ 3.13.0, we started to use the blog to publish
announcements, so they are not needed anymore.

The `versions` page was cleaned up. It now show the release branches
and their state of support in the same way as Linux distribution
versions on Wikipedia [1]. It is now possible for a user to see what is
supported at a glance.

The really old releases were deleted from the page. We don't need to
know that RabbitMQ 3.1.x is no longer supported.

The `versions.md` page was renamed to `index.md`. Thus the new URL is
simply `/release-information`.

All release details are put in `docusaurus.config.js`. This was already
used to know what the latest patch was on a per-release-branch basis.
Now, we also set the release and end-of-support dates.

They are rendered as a table using a React/MDX component in
`src/components/RabbitMQServerReleaseInfo`. The component selects the
colors based on the compile-time date and the end-of-support dates.
Therefore, the colors will be automatically updated with each build
of the website. This is not rendered at runtime, so if there are no
builds of the website, the color will be outdated. I believe that's ok
because support expires at the same time a new release series is
published.

[1] https://en.wikipedia.org/wiki/Debian_version_history