Version images with the docs
[Why]
Currently, images are all in a global directory, `static/img`.

Docusaurus allows to colocate images (and files in general) with the doc
that uses them. This allows to version them alongside the doc itself.
Thus, if we need to update an image for a new version of RabbitMQ, we
can do that without affecting older versions.

[How]
To have other files "attached" to a document, it has to be a directory.
The document is named `index.md` inside this directory. For example:

    docs/feature-flags/index.md

The URL remains the same; `/docs/feature-flags` in this example.

Then images are moves to that directory:

    docs/feature-flags/index.md
    docs/feature-flags/image.svg

Finally, images are referenced relatively to `index.md`:

    ![](./image.svg)

All unused images were deleted in the process.