Remove some HTML tables
As I'm working on fixing their content as part of the previous commit
about `RabbitMQServer*` components and the next 3.13 docs version, let's
use Markdown tables instead of HTML tables. Their layout is better and
consistent with the rest of the docs.

At some point, we should use another layout for this informations as the
table is not the appropriate tool here.