Tutorials: Clarify they work with any versions of RabbitMQ
[Why]
Now that the docs are versioned and that the currently browsed version
is displayed at the top of versioned pages, we need some clarification
in the non-versioned pages.

The Release information page already talks about all versions of
RabbitMQ, so we are fine there.

The Kubernetes Operator docs are about the operator, not RabbitMQ, so we
should be ok too. That said, it could be valuable to clarify that point,
but I'm not familiar with this section.

Finally, the tutorials. They work with any versions of RabbitMQ and we
can be clear about this. That's what this commit is about.