release-information: List all patch releases
[Why]
With the switch to the old news+changelog pages to the new release
information one, we lost the list of patch releases with their release
date.

This commit restores the list of patch releases.

[How]
To not bloat the screen/page, only the latest patch for each release
branch is displayed.

There is an icon to let the user unfold the list and show all patch
releases for a given release branch.

V2: The initial patch was only about the new support policy and
    especially the change of end of community support date change. This
    new patch keeps that but also lists all patch release for each
    release branch.

V3: Support small screens, with CSS tunings and new SVG icons.