cloudflare-bulk-redirects.csv: Add redirect `/docs/3.13` -> `/docs`
[Why]
This will allow us to point to the 3.13 docs in external resources like
the RabbitMQ management UI before we cut the next release.

For instance, the management UI could create a link such as
`/docs/3.13/feature-flags`. While RabbitMQ 3.13.x remains the latest
release branch, it will redirect to `/docs/feature-flags`. However, once
we cut the next release branch in the docs (for example 3.14), we can
remove the redirect and the link will point to the real
`/docs/3.13/feature-flags` (while `/docs/feature-flags` corresponds to
3.14).