Render blog images with a caption.
The Markdown syntax is the following:

```
![this will go in both alt and caption](path)
![this will go in alt|this will go in caption](path)
![alt only|](path)
```

This is handled by a Rehype plugin.

The goal is to have captions for the images in blog posts, as it
was in the previous blog. Most blog posts need a caption for their
images as the text either refers to the images or the caption contains
a summary of the surrounding, which makes the post easier to "browse".