Add documentation about Khepri
[Why]
Khepri is a large addition of RabbitMQ 4.0.0. It is invisible from
outside but greatly affects the behavior of RabbitMQ once enabled.

Furthermore, it is important to explain what should users expect, to let
them make an informed decision about if/when Khepri should be enabled.

[How]
The patch adds a new "Metadata store" section under "How to manage
RabbitMQ". It uses the word metadata store because it explains the role
of that subsystem of RabbitMQ as well as the difference between Mnesia
and Khepri.

It also adds pages specific to Khepri:
* how to enable it
* how it affects clustering
* how it affects daily use for application developers
* how it handles failure recovery
* the known issues
* frequently asked questions