plugins/configure-svgo.js: Docusaurus plugin to fix SVG inlining
[Why]
In Docusaurus, it is possible to inline SVG by importing them:

    import MetadataStoreCluster from './metadata-store-cluster.svg';

    <MetadataStoreCluster/>

This allows to override the styling of the SVG. This is really useful to
make SVG which are compatible with the light and dark themes at the same
time. Here is an example of a CSS snippet overriding the style for the
imported SVG above for the dark theme:

    .diagram tspan[fill="#000000"] {
      fill: var(--docsearch-text-color) !important;
    }

    .diagram rect[stroke="#000000"],
    .diagram path[stroke="#000000"] {
      stroke: var(--docsearch-text-color) !important;
    }

In this example, we set the `fill` of text elements and the `stroke` of
rectangles and paths that are black to the same color as the text in the
page.

Unfortunalety, Docusaurus' dependency `svgo`, which wants to optimize
SVG just break them, thowing some styling internal to the SVG away.

[How]
The plugin added by this commit overrides the configuration of `svgo` to
only allow a small list of features to make the SVG compatible with JSX
components and that's it.

In the end, inlined SVG are mostly vanilla and they are displayed and
styled as expected.