Feature flags: update Khepri description for 4.0
This flag is very vast in scope and for that reason,
still opt-in at the moment.

However, it has matured and should be considered
for production use as of 4.0, after an adequeate
period of testing in non-production environments.

Some users already find Khepri a more suitable
(in terms of failure recovery) schema data store
compared to Mnesia.

It obvious has fewer collective community mileage
than Mnesia (which has been used in RabbitMQ for
some 17-and-a-half years since its inception in 2007)
but the only way to test it heavily is to recommend
that the community does so.

Per discussion with the team.