Add a "What's new" page for RabbitMQ 4.0
[Why]
This is expected by the release information page.

[How]
The page is largely inspired by the 3.13 equivalent and the RabbitMQ
4.0.1 release notes.

I reproduced the AMQP SVG logo for this occasion because there is only
an blurry PNG version on the official AMQP website.

Logo style are fixed in a few places as well.

The metadata store docs now uses the inline Khepri SVG logo.