Update Event Exchange Plugin docs
This is the docs PR for
https://github.com/rabbitmq/rabbitmq-server/pull/12714 and applies to
RabbitMQ >= 4.1

Note that the internal event types were documented previously at three
different places:
1. https://github.com/rabbitmq/rabbitmq-server/blob/main/deps/rabbitmq_event_exchange/README.md
2. https://www.rabbitmq.com/docs/event-exchange
3. https://www.rabbitmq.com/docs/logging#internal-events

All of them listed different and/or outdated event types.
From now on, we use the 3rd place as only place and single source of
truth for documenting internal events.