Move AMQP 0-9-1 reference to a GitHub repo
The primary reason for this change is: at the
moment, one of the Docusaurus/React/MDX
dependencies does not handle large HTML
pages very well and runs into a stack overflow
of Promise calls.

This has proven to be a really costly problem
to debug, and seemingly fairly unique in the
large community of Docusaurus users.

Large static HTML tables are just not that common
for a lot of OSS software, I guess.

Given that the AMQP 0-9-1 spec
is frozen in time and extensions are added very
infrequently, it was decided to move it to
GitHub and also archive the PDFs and
the XML files developed for code generation
there.

The Quick Reference, the concept page,
the list of extensions and the specification
pages are significantly smaller and do not
cause any pains during rendering, so we can copy
them to the GitHub repo and continue serving
them as part of the site.