DQT-related doc section updates
Since node-wide DQT in 3.13.x is not really
fully supported (it only applies to virtual
hosts that have some metadata configured),
let's not mention that option in the 3.13.x
version of the guide.

As for 4.0.x and 4.1.x, they don't have to mention
that this feature or that first shipped in
a `3.13.x` patch release.