Add note about interaction of queue mirroring and anonymous queues
We had an interesting incident because our anonymous queues, although declared with 

```
Durable:            false,
AutoDelete:         true,
```

and a message ttl, were not deleted when their consumer was cancelled. Each pod in our kubernetes deployment created an anonymous queue, and then failed to clean it up. As a result we saw disk utilization alarms and had to take emergency action. 

It turns out that a) they were not declared with exlusive (which according to docs would have turned off mirroring) and b) they had mirroring applied with a policy, which apparently turned off auto-deletion.

I know mirroring is deprecated, which I called out in this PR, but I'm sure many deployments, like us, have not yet switched to quorum queues, so I think some documentation support is warranted.