New guide: HTTP API reference
This is a reformatted and edited version of
the same page served by the HTTP API on
every node the management plugin is enabled on.

However, this version is arguably easier
to follow and edit, in particular add examples
to, without being nearly as constraint
by the available horizontal space.