Fix undefined method call: replace body() with getBody() in rpc_client.php
Replaced the incorrect call to AMQPMessage::body() with the correct method AMQPMessage::getBody() in rpc_client.php.

The 'body()' method does not exist in PhpAmqpLib\Message\AMQPMessage, leading to a fatal error at runtime. According to the PhpAmqpLib documentation, 'getBody()' should be used to retrieve the message content.

This change ensures compatibility with the current version of php-amqplib/php-amqplib and prevents runtime errors during RPC calls.