Blue-Green Upgrade guide updates
rabbitmqadmin v2 provides a new set of commands
that greatly simplify the task of enabling
queue federation of all policies.

To make this update possible for the 3.13
edition, this backports some Policies
guide changes to that edition.
After all, the rabbitmqadmin v2 commands
in question can be used with 3.13.x nodes
as well as 4.x ones.