Debian and Ubuntu guide: the end of the Cloudsmith era
Cloudsmith is a solid service that's served us
for nearly a decade.

However, we have outgrown them years ago,
and had to host our own mirrors with close to
two orders of magnitude more bandwidth available.

Now we have reasons to migrate entirely to our
own repositories, for a number of reasons
ranging from the cost efficiency at our scale
to the need to reduce external service dependencies.

Thank you, the Cloudsmith team, for building
a great service and being a good ally of RabbitMQ.