Mention encrypted listeners in protocol-listener API reference
The protocol-listener health check accepts protocol names with 's' for
the encrypted listeners like `amqps` and `mqtts`. The health check
distinguishes between encrypted and non-encrypted so to get a 200 when
running AMQPS but not AMQP you need to query for `amqps`.

This change mentions the encrypted protocol names. I've also combined
`amqp091` and `amqp10` as `amqp` which is what the health check
translates these into.