Document the name of the credentials Secret generated for User objects
Knowing the Secret name before it is generated is required for some
workflows.

For instance, if you need to refer to the Secret to mount it in a Pod,
and you're not deploying your User object before writing your Pod spec
(common in gitops workflows), you need to have the Secret name
beforehand.

The Secret name is already fixed in code, but this is currently not
documented.

Update the documenation in accordance with the code.