RPM guide: explain how often the mirrors are synced
Due to certain reasons about how the mirrors were
managed historically, the RPM one now can be synced at
certain moments during the month.

In order to lift this restriction the way we did
for its Debian counterpart, we need to create a new
repo, import every package from the current repo there
and re-sign all packages and the repo at the end.

That's doable and was done for Debian but it was
fairly disruptive and took months, plus hours spent
debugging installation issues that had nothing to
do with the mirrors or our tooling used to manage them.

So for RPM packages, the idea is to wait and see
if we need to migrate the repo or this syncing
interval would be acceptable for the (generally)
slower moving RPM-based distribution community.